+++ modulo3/projeto-labefy/projeto-labefy/src/components/PlayListScreen.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+
+class App extends React.Component{
+    
+}
